backend/requirements.txt
anthropic>=0.84.0
python-docx>=1.2.0
boto3>=1.42.0
fastapi>=0.134.0
uvicorn>=0.41.0
python-multipart>=0.0.22
pydantic>=2.12.0
python-dotenv>=1.2.0
aiofiles>=25.1.0
